+++ packages/plugin/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./plugin"
 export * from "./persist"
 export * from "./addr"
+export type { TxCreator, TxArgSpec } from "@polkadot-api/tx-creator"