+++ plugins/multisig/src/MultisigExternalSigner.tsx
@@ -1,35 +1,40 @@
+import { getMultisigTxCreator } from "@polkadot-api/meta-signers"
 import {
   AccountId,
-  Binary,
   getMultisigAccountId,
   HexString,
 } from "@polkadot-api/substrate-bindings"
-import { CreateMultisigSigner, MultisigInfo } from "./provider"
-import { createSignal } from "@react-rxjs/utils"
-import { state, useStateObservable } from "@react-rxjs/core"
-import { filter, firstValueFrom } from "rxjs"
-import { FC } from "react"
+import type { TxCreator } from "@polkahub/plugin"
 import {
   Dialog,
+  DialogBody,
   DialogContent,
   DialogHeader,
   DialogTitle,
-  DialogBody,
 } from "@polkahub/ui-components"
+import { state, useStateObservable } from "@react-rxjs/core"
+import { createSignal } from "@react-rxjs/utils"
 import { Link } from "lucide-react"
+import { FC } from "react"
+import { filter, firstValueFrom } from "rxjs"
+import {
+  CreateMultisigTxCreator,
+  IdentifiedTxCreator,
+  MultisigInfo,
+} from "./provider"
 
 const [urlChange$, setUrl] = createSignal<string | null>()
 const url$ = state(urlChange$, null)
 
 const [enc] = AccountId()
 export const multisigExternalSigner =
-  (
+  <T extends IdentifiedTxCreator>(
     getMultisigUrl: (
       info: MultisigInfo,
       callData: HexString,
     ) => string | Promise<string>,
-    thresholdOneFallback?: CreateMultisigSigner,
-  ): CreateMultisigSigner =>
+    thresholdOneFallback?: CreateMultisigTxCreator<T>,
+  ): CreateMultisigTxCreator<T> =>
   (info, signer) => {
     if (info.threshold === 1 && signer && thresholdOneFallback)
       return thresholdOneFallback(info, signer)
@@ -39,22 +44,32 @@ export const multisigExternalSigner =
       signatories: info.signatories.map(enc),
     })
 
-    return {
-      publicKey,
-      signBytes() {
-        throw new Error("Raw bytes can't be signed with a multisig")
-      },
-      async signTx(callData) {
-        const url = await getMultisigUrl(info, Binary.toHex(callData))
-        setUrl(url)
-        try {
-          await firstValueFrom(url$.pipe(filter((v) => !v)))
-          throw null
-        } catch (ex) {
-          throw new Error("Dismissed")
-        }
-      },
+    const creator: TxCreator = async (txPayload, opts, bindings, fake) => {
+      const { callData } = txPayload
+      if (fake) {
+        if (!signer) throw new Error("Needs a parent signer to fake sign")
+
+        const multisigSigner = getMultisigTxCreator(
+          info,
+          async () => undefined,
+          signer,
+        )
+        return multisigSigner(txPayload, opts, bindings, true)
+      }
+      const url = await getMultisigUrl(info, callData)
+      setUrl(url)
+      try {
+        await firstValueFrom(url$.pipe(filter((v) => !v)))
+        throw null
+      } catch (ex) {
+        throw new Error("Dismissed")
+      }
     }
+
+    return Object.assign(creator as T, {
+      accountId: publicKey,
+      publicKey,
+    })
   }
 
 export const MultisigExternalSignerModal: FC = () => {