+++ plugins/read-only/src/provider.ts
@@ -7,8 +7,7 @@ import {
   Plugin,
 } from "@polkahub/plugin"
 import { DefaultedStateObservable, withDefault } from "@react-rxjs/core"
-import { AccountId, Binary } from "polkadot-api"
-import { getPolkadotSigner } from "polkadot-api/signer"
+import { getFakeTxCreator, RawTxCreator } from "polkadot-api/tx-creator"
 import { map } from "rxjs"
 
 export interface ReadonlyAccountInfo {
@@ -17,13 +16,14 @@ export interface ReadonlyAccountInfo {
 }
 
 export const readOnlyProviderId = "readonly"
-export interface ReadOnlyProvider extends Plugin {
+type ReadOnlyAccount = Account<RawTxCreator>
+export interface ReadOnlyProvider extends Plugin<ReadOnlyAccount> {
   id: "readonly"
-  accounts$: DefaultedStateObservable<Account[]>
+  accounts$: DefaultedStateObservable<ReadOnlyAccount[]>
   setAccounts: (payload: ReadonlyAccountInfo[]) => void
-  addAccount: (address: ReadonlyAccountInfo) => Account
+  addAccount: (address: ReadonlyAccountInfo) => ReadOnlyAccount
   removeAccount: (address: AccountAddress) => void
-  toAccount: (address: AccountAddress) => Account
+  toAccount: (address: AccountAddress) => ReadOnlyAccount
 }
 
 export const createReadOnlyProvider = (
@@ -51,11 +51,14 @@ export const createReadOnlyProvider = (
         }
       : value
 
-  const getAccount = ({ address, name }: ReadonlyAccountInfo): Account => ({
+  const getAccount = ({
+    address,
+    name,
+  }: ReadonlyAccountInfo): ReadOnlyAccount => ({
     name,
     provider: readOnlyProviderId,
     address,
-    signer: fakeSigner ? createFakeSigner(address) : undefined,
+    txCreator: fakeSigner ? getFakeTxCreator(address) : undefined,
   })
 
   const accounts$ = persistedAccounts$.pipeState(
@@ -91,18 +94,3 @@ export const createReadOnlyProvider = (
       }),
   }
 }
-
-const createFakeSigner = (address: AccountAddress) =>
-  getPolkadotSigner(
-    address.startsWith("0x")
-      ? Binary.fromHex(address)
-      : AccountId().enc(address)!,
-    "Sr25519",
-    () => {
-      // From https://wiki.acala.network/build/sdks/homa
-      const signature = new Uint8Array(64)
-      signature.fill(0xcd)
-      signature.set([0xde, 0xad, 0xbe, 0xef])
-      return signature
-    },
-  )